another fix due to settings changes

+++ dao/src/test/java/it/com/inthinc/pro/dao/SiloServiceTest.java
@@ -301,7 +301,7 @@ public class SiloServiceTest {
         assertTrue(settings.size()==2);
         assertTrue(settings.get(1234) != null);
         sensitivitySliderValues = settings.get(1234);
-        assertEquals(new Integer(3),(Integer)sensitivitySliderValues.getDefaultValueIndex());
+        assertEquals(new Integer(8),(Integer)sensitivitySliderValues.getDefaultValueIndex());
         
         settingValues = slider.getSettingValuesFromSliderValue(1);
         assertTrue(settingValues.size()==2);
